[SOL] LLDB for LLVM 19 (#145)
* [SOL] lldb: pretty print instructions

* [SOL] lldb: skip ranges with DW_AT_low_pc equal to zero

* [SOL] lldb: Add solana-lldb wrapper

This adds a wrapper around the LLDB executable that resides in the same directory. Once we start to distribute LLDB in sbf-tools, the wrapper + scripts can be put into the same directory as the LLDB executable. Invoking ./solana-lldb will pre-load helper functions to print rust as well as solana types (e.g. pubkeys in base58 or auto deref of accounts) and then start the LLDB executable.

* [SOL] Revamp LLDB for SBF

---------

Co-authored-by: wj <wzggUaPFt5J7f@gmx.de>